requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
aiomysql
langchain
openai
faiss-cpu
httpx
pydantic
pytest
faker
